move some components functions to base node
