+++ cocos2d/core/utils/base-node.js
@@ -28,6 +28,7 @@ var Flags = require('../platform/CCObject').Flags;
 var Misc = require('./misc');
 var IdGenerater = require('../platform/id-generater');
 
+var JS = cc.js;
 var Destroying = Flags.Destroying;
 var DontDestroy = Flags.DontDestroy;
 var Activating = Flags.Activating;
@@ -272,8 +273,11 @@ var BaseNode = cc.Class(/** @lends cc.Node# */{
 
     },
 
-    ctor: function () {
+    ctor: function (name) {
 
+        this._name = typeof name !== 'undefined' ? name : 'New Node';
+
+        this._activeInHierarchy = false;
         // Support for ActionManager and EventManager
         this.__instanceId = this._id || cc.ClassManager.getNewInstanceId();
 
@@ -468,6 +472,305 @@ var BaseNode = cc.Class(/** @lends cc.Node# */{
         return false;
     },
 
+    // COMPONENT
+
+    /**
+     * !#en
+     * Returns the component of supplied type if the node has one attached, null if it doesn't.<br/>
+     * You can also get component in the node by passing in the name of the script.
+     * !#zh
+     * 获取节点上指定类型的组件，如果节点有附加指定类型的组件，则返回，如果没有则为空。<br/>
+     * 传入参数也可以是脚本的名称。
+     * @method getComponent
+     * @param {Function|String} typeOrClassName
+     * @return {Component}
+     * @example
+     * // get sprite component.
+     * var sprite = node.getComponent(cc.Sprite);
+     * // get custom test calss.
+     * var test = node.getComponent("Test");
+     */
+    getComponent: function (typeOrClassName) {
+        var constructor = getConstructor(typeOrClassName);
+        if (constructor) {
+            return findComponent(this, constructor);
+        }
+        return null;
+    },
+
+    /**
+     * !#en Returns all components of supplied type in the node.
+     * !#zh 返回节点上指定类型的所有组件。
+     * @method getComponents
+     * @param {Function|String} typeOrClassName
+     * @return {Component[]}
+     * @example
+     * var sprites = node.getComponents(cc.Sprite);
+     * var tests = node.getComponents("Test");
+     */
+    getComponents: function (typeOrClassName) {
+        var constructor = getConstructor(typeOrClassName), components = [];
+        if (constructor) {
+            findComponents(this, constructor, components);
+        }
+        return components;
+    },
+
+    /**
+     * !#en Returns the component of supplied type in any of its children using depth first search.
+     * !#zh 递归查找所有子节点中第一个匹配指定类型的组件。
+     * @method getComponentInChildren
+     * @param {Function|String} typeOrClassName
+     * @return {Component}
+     * @example
+     * var sprite = node.getComponentInChildren(cc.Sprite);
+     * var Test = node.getComponentInChildren("Test");
+     */
+    getComponentInChildren: function (typeOrClassName) {
+        var constructor = getConstructor(typeOrClassName);
+        if (constructor) {
+            return findChildComponent(this._children, constructor);
+        }
+        return null;
+    },
+
+    /**
+     * !#en Returns all components of supplied type in self or any of its children.
+     * !#zh 递归查找自身或所有子节点中指定类型的组件
+     * @method getComponentsInChildren
+     * @param {Function|String} typeOrClassName
+     * @return {Component[]}
+     * @example
+     * var sprites = node.getComponentsInChildren(cc.Sprite);
+     * var tests = node.getComponentsInChildren("Test");
+     */
+    getComponentsInChildren: function (typeOrClassName) {
+        var constructor = getConstructor(typeOrClassName), components = [];
+        if (constructor) {
+            findComponents(this, constructor, components);
+            findChildComponents(this._children, constructor, components);
+        }
+        return components;
+    },
+
+    _checkMultipleComp: CC_EDITOR && function (ctor) {
+        var existing = this.getComponent(ctor._disallowMultiple);
+        if (existing) {
+            if (existing.constructor === ctor) {
+                cc.errorID(3805, JS.getClassName(ctor), this._name);
+            }
+            else {
+                cc.errorID(3806, JS.getClassName(ctor), this._name, JS.getClassName(existing));
+            }
+            return false;
+        }
+        return true;
+    },
+
+    /**
+     * !#en Adds a component class to the node. You can also add component to node by passing in the name of the script.
+     * !#zh 向节点添加一个指定类型的组件类，你还可以通过传入脚本的名称来添加组件。
+     * @method addComponent
+     * @param {Function|String} typeOrClassName - The constructor or the class name of the component to add
+     * @return {Component} - The newly added component
+     * @example
+     * var sprite = node.addComponent(cc.Sprite);
+     * var test = node.addComponent("Test");
+     */
+    addComponent: function (typeOrClassName) {
+
+        if (CC_EDITOR && (this._objFlags & Destroying)) {
+            cc.error('isDestroying');
+            return null;
+        }
+
+        // get component
+
+        var constructor;
+        if (typeof typeOrClassName === 'string') {
+            constructor = JS.getClassByName(typeOrClassName);
+            if ( !constructor ) {
+                cc.errorID(3807, typeOrClassName);
+                if (cc._RFpeek()) {
+                    cc.errorID(3808, typeOrClassName);
+                }
+                return null;
+            }
+        }
+        else {
+            if ( !typeOrClassName ) {
+                cc.errorID(3804);
+                return null;
+            }
+            constructor = typeOrClassName;
+        }
+
+        // check component
+
+        if (typeof constructor !== 'function') {
+            cc.errorID(3809);
+            return null;
+        }
+        if (!cc.isChildClassOf(constructor, cc.Component)) {
+            cc.errorID(3810);
+            return null;
+        }
+
+        if (CC_EDITOR && constructor._disallowMultiple) {
+            if (!this._checkMultipleComp(constructor)) {
+                return null;
+            }
+        }
+
+        // check requirement
+
+        var ReqComp = constructor._requireComponent;
+        if (ReqComp && !this.getComponent(ReqComp)) {
+            var depended = this.addComponent(ReqComp);
+            if (!depended) {
+                // depend conflicts
+                return null;
+            }
+        }
+
+        //// check conflict
+        //
+        //if (CC_EDITOR && !_Scene.DetectConflict.beforeAddComponent(this, constructor)) {
+        //    return null;
+        //}
+
+        //
+
+        var component = new constructor();
+        component.node = this;
+        this._components.push(component);
+
+        if (this._activeInHierarchy) {
+            if (typeof component.__preload === 'function') {
+                cc.Component._callPreloadOnComponent(component);
+            }
+            // call onLoad/onEnable
+            component.__onNodeActivated(true);
+        }
+
+        return component;
+    },
+
+    /**
+     * This api should only used by undo system
+     * @method _addComponentAt
+     * @param {Component} comp
+     * @param {Number} index
+     * @private
+     */
+    _addComponentAt: CC_EDITOR && function (comp, index) {
+        if (this._objFlags & Destroying) {
+            return cc.error('isDestroying');
+        }
+        if ( !(comp instanceof cc.Component) ) {
+            return cc.errorID(3811);
+        }
+        if (index > this._components.length) {
+            return cc.errorID(3812);
+        }
+
+        // recheck attributes because script may changed
+        var ctor = comp.constructor;
+        if (ctor._disallowMultiple) {
+            if (!this._checkMultipleComp(ctor)) {
+                return;
+            }
+        }
+        var ReqComp = ctor._requireComponent;
+        if (ReqComp && !this.getComponent(ReqComp)) {
+            if (index === this._components.length) {
+                // If comp should be last component, increase the index because required component added
+                ++index;
+            }
+            var depended = this.addComponent(ReqComp);
+            if (!depended) {
+                // depend conflicts
+                return null;
+            }
+        }
+
+        comp.node = this;
+        this._components.splice(index, 0, comp);
+
+        if (this._activeInHierarchy) {
+            if (typeof comp.__preload === 'function' &&
+                !(comp._objFlags & cc.Object.Flags.IsPreloadCalled)) {
+                cc.Component._callPreloadOnComponent(comp);
+            }
+            // call onLoad/onEnable
+            comp.__onNodeActivated(true);
+        }
+    },
+
+    /**
+     * !#en
+     * Removes a component identified by the given name or removes the component object given.
+     * You can also use component.destroy() if you already have the reference.
+     * !#zh
+     * 删除节点上的指定组件，传入参数可以是一个组件构造函数或组件名，也可以是已经获得的组件引用。
+     * 如果你已经获得组件引用，你也可以直接调用 component.destroy()
+     * @method removeComponent
+     * @param {String|Function|Component} component - The need remove component.
+     * @deprecated please destroy the component to remove it.
+     * @example
+     * node.removeComponent(cc.Sprite);
+     * var Test = require("Test");
+     * node.removeComponent(Test);
+     */
+    removeComponent: function (component) {
+        if ( !component ) {
+            cc.errorID(3813);
+            return;
+        }
+        if (typeof component !== 'object') {
+            component = this.getComponent(component);
+        }
+        if (component) {
+            component.destroy();
+        }
+    },
+
+    /**
+     * @method _getDependComponent
+     * @param {Component} depended
+     * @return {Component}
+     * @private
+     */
+    _getDependComponent: CC_EDITOR && function (depended) {
+        for (var i = 0; i < this._components.length; i++) {
+            var comp = this._components[i];
+            if (comp !== depended && comp.isValid && !cc.Object._willDestroy(comp)) {
+                var depend = comp.constructor._requireComponent;
+                if (depend && depended instanceof depend) {
+                    return comp;
+                }
+            }
+        }
+        return null;
+    },
+
+    // do remove component, only used internally
+    _removeComponent: function (component) {
+        if (!component) {
+            cc.errorID(3814);
+            return;
+        }
+
+        if (!(this._objFlags & Destroying)) {
+            var i = this._components.indexOf(component);
+            if (i !== -1) {
+                this._components.splice(i, 1);
+            }
+            else if (component.node !== this) {
+                cc.errorID(3815);
+            }
+        }
+    },
 });
